requirements.txt
# XClaw Dependencies
# ─────────────────────────────────────────────────────────────────────────────
# Core (required)
pyyaml>=6.0

# Web interface
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# Telegram interface
python-telegram-bot>=21.0

# .env loading
python-dotenv>=1.0.0

# Optional: for richer web scraping in research agent
# beautifulsoup4>=4.12.0
# httpx>=0.27.0
